Added functional tests, Changelog file, refactored tests and downgraded Qodana action
