+++ .github/workflows/build.yml
@@ -118,6 +118,6 @@ jobs:
 
       # Run Qodana inspections
       - name: Qodana - Code Inspection
-        uses: JetBrains/qodana-action@v2024.1.8
+        uses: JetBrains/qodana-action@v2023.2.8
         with:
           cache-default-branch-only: true
