+++ CHANGELOG.md
@@ -1,0 +1,10 @@
+<!-- Keep a Changelog guide -> https://keepachangelog.com -->
+
+# Properties Manager Changelog
+
+## [Unreleased]
+### Added
+- Added propertyManager extension.
+- Added functions to bind single value properties for the following types: String, Char, Int, Long, Float, Double, BigInteger, BigDecimal, Boolean.
+- Added function to bind single value properties for the custom types.
+- Added function to bind multi value properties.
