+++ README.md
@@ -1,3 +1,5 @@
 # properties-manager
 
-Plugin to load and bind properties
+Plugin to load and bind properties.
+
+## Examples:
